Add CMAC and helper instruction
CMAC is a continous multiply-accumulate instruction. It is setup using
special instructions that write a coefficient address, a data addresse,
a data offset and a data count to special registers. These initial
values are then iterated through, multiplying each corresponding value
in the data and coefficient arrays and accumulating. The final result is
stored in the Accumulator. CMAC stalls the program counter for
4+data_size cycles.